refactor(DropdownTrigger): accept startContent instead of tokens
