+++ src/DropdownTrigger/index.stories.js
@@ -2,6 +2,7 @@ import React, { useState } from "react";
 import DropdownTrigger from "./";
 import Popover from "../Popover";
 import RadioButtons from "../RadioButtons";
+import FieldToken from "../FieldToken";
 
 const Template = (args) => <DropdownTrigger {...args} />;
 
@@ -20,7 +21,7 @@ LabelOnly.parameters = {
   },
 };
 
-export const WithTokens = () => {
+export const WithTokensAsStartContent = () => {
   const [tokens, setTokens] = useState([
     "Detroit, MI",
     "Chicago, IL",
@@ -31,17 +32,37 @@ export const WithTokens = () => {
     "Houston, TX",
     "Philadelphia, PA",
   ]);
+
+  const handleTokenDismiss = (token) => {
+    setTokens((oldTokens) => {
+      const newTokens = new Set(oldTokens);
+      newTokens.delete(token);
+      return [...newTokens];
+    });
+  };
+
   return (
     <DropdownTrigger
-      tokens={tokens}
-      onTokensChange={(tokes) => setTokens(tokes)}
+      displayValue={null}
+      startContent={
+        <span className="nds-dropdownTrigger-tokens padding--y--xs">
+          {tokens.map((label) => (
+            <FieldToken
+              key="label"
+              label={label}
+              onDismiss={handleTokenDismiss}
+            />
+          ))}
+        </span>
+      }
     />
   );
 };
-WithTokens.parameters = {
+WithTokensAsStartContent.parameters = {
   docs: {
     description: {
-      story: "Tokens in `DropdownTrigger` are fully controlled.",
+      story:
+        "The `startContent` and `endContent` props may be used to render arbitrary JSX, like tokens and icons.",
     },
   },
 };
